requirements.txt
fastapi>=0.104.1
matplotlib>=3.8.0
numpy>=1.26.2
pandas>=2.1.3
Pillow>=10.1.0
pydicom>=2.4.3
